controller: Improve overlay event emission.
This patch introduces an enum for openvr-glib controller button types,
which can be mapped to OpenVR EVRButtonId.
It also uses a common function to check for press / release events and
eliminates the use of gbooleans per button as OpenVRController member
and caches the OpenVR state in an uint64_t instead.
It also names overlay related polling functions more explicitly.