openvr-context: Adapt to OpenVR 1.8.19 scene transition changes
Scene applications do not get VREvent_SceneApplicationStateChanged
when another scene app starts. They only get a VREvent_Quit.
At this point they do not know whether they got this quit event because
the user exited SteamVR, or because the user started another scene app.

xrdesktop for example wants to know this difference because if it runs in
scene mode and the user starts another scene app, xrdesktop wants to transition
to overlay mode and if the user quits steamvr, xrdesktop wants to quit.

However when scene apps get a VREvent_Quit they can look at the SceneApplicationState.
If this state is "quitting" that seems to mean steamvr is trying to shut us down, in order
to start another scene app.
If this state is "running", that seems to mean steamvr is shutting down itself.