context: consolidate openxr and openvr initialization
For a common initialization order the initialization had to be split:

OpenVR:
1. Init OpenVR runtime
2. Get Vulkan extension strings from OpenVR, init gulkan with them.
   Unfortunately must happen after 1., would be nice to have a lightweight function for it.
   https://github.com/ValveSoftware/openvr/issues/993
3. Nothing

OpenXR:
1. Init OpenXR Instance, System etc. (everything that doesn't require vulkan)
2. Init gulkan (generic, doesn't need anything from XR runtime)
3. Init graphics binding, session (requires vulkan)

With this split gxr can also support applications that want to use the VR runtime, but not
render to it, though OpenXR headless sessions are not yet implemented in gxr.

Most tests/examples will still fail on OpenXR at this point.