context: add request_quit()
In OpenVR, an application that wants to quit just does it,
and leaves OpenVR to deal with transitions to other applications.

In OpenXR, an application requests the runtime to quit, and then
keeps running until the runtime changes its state to STOPPING (and notifies
the application with an event).

Unified quit procedure:
An application that wants to quit calls
    gxr_context_request_quit()
and then waits for a quit signal from the GxrContext.
OpenVRContext sends it immediately.
OpenXRContext sends it when the runtime is in STOPPING state.
When handling the quit sgnale, the application calls
    gxr_context_acknowledge_quit()
which is necessary on OpenVR, but a noop on OpenXR.