examples/cube: Pack all ub structs
don't take address of member of packed struct

Pointers to unaligned memory are not guaranteed to be handled correctly when passed to functions that are unaware of alignment.

For example in code like

typedef struct __attribute__((__packed__)) {
...
  float arr[2];
...
} S;
int main()
{
  S s;
  library_function(s.arr);
}

library_function() is not aware that s.arr might not start at an address aligned to the platform's convention.

The performance bottleneck is certainly not here, so just copy values individually.