openvr-overlay: Don't ref gxr context
xrdesktop uses overlays for the pointer tip and pointer.
When the openvr context is destroyed, the device manager is destroyed,
which destroys the controllers including pointers and tips.
Triggering this chain by destroying the openvr context is not possible
while the pointer and tip ref the openvr context.

It's the app's responsibility to destroy all overlays before the gxr context.