Remove GXR_QUIT_PROCESS_QUIT
On OpenVR we were notified by a quit signal that a scene application quit
while we were running as overlay application.
On OpenXR we always run as "Overlay" application (or non overlay if the
runtime doesn't support it), We still have the signal that a main
application quit, but it is now an OverlayEvent and separated from the
quit events.